add sym network factory
