+++ sym_demo.py
@@ -10,6 +10,7 @@ from symdata.loader import load_test, generate_batch
 from symdata.vis import vis_detection
 from symimdb.dataset import get_dataset_demo
 from symnet.model import load_param, check_shape, get_max_shape_test
+from symnet.network import get_network_test
 
 
 def demo_net(sym, class_names, args):
@@ -108,68 +109,10 @@ def parse_args():
     return args
 
 
-def get_vgg16_test(args):
-    from symnet.symbol_vgg import get_vgg_test
-    args.img_pixel_means = (123.68, 116.779, 103.939)
-    args.img_pixel_stds = (1.0, 1.0, 1.0)
-    args.net_fixed_params = ['conv1', 'conv2']
-    args.rpn_feat_stride = 16
-    args.rcnn_feat_stride = 16
-    args.rcnn_pooled_size = (7, 7)
-    return get_vgg_test(anchor_scales=args.rpn_anchor_scales, anchor_ratios=args.rpn_anchor_ratios,
-                        rpn_feature_stride=args.rpn_feat_stride, rpn_pre_topk=args.rpn_pre_nms_topk,
-                        rpn_post_topk=args.rpn_post_nms_topk, rpn_nms_thresh=args.rpn_nms_thresh,
-                        rpn_min_size=args.rpn_min_size,
-                        num_classes=args.rcnn_num_classes, rcnn_feature_stride=args.rcnn_feat_stride,
-                        rcnn_pooled_size=args.rcnn_pooled_size, rcnn_batch_size=args.rcnn_batch_size)
-
-
-def get_resnet50_test(args):
-    from symnet.symbol_resnet import get_resnet_test
-    args.img_pixel_means = (0.0, 0.0, 0.0)
-    args.img_pixel_stds = (1.0, 1.0, 1.0)
-    args.rpn_feat_stride = 16
-    args.rcnn_feat_stride = 16
-    args.rcnn_pooled_size = (14, 14)
-    return get_resnet_test(anchor_scales=args.rpn_anchor_scales, anchor_ratios=args.rpn_anchor_ratios,
-                           rpn_feature_stride=args.rpn_feat_stride, rpn_pre_topk=args.rpn_pre_nms_topk,
-                           rpn_post_topk=args.rpn_post_nms_topk, rpn_nms_thresh=args.rpn_nms_thresh,
-                           rpn_min_size=args.rpn_min_size,
-                           num_classes=args.rcnn_num_classes, rcnn_feature_stride=args.rcnn_feat_stride,
-                           rcnn_pooled_size=args.rcnn_pooled_size, rcnn_batch_size=args.rcnn_batch_size,
-                           units=(3, 4, 6, 3), filter_list=(256, 512, 1024, 2048))
-
-
-def get_resnet101_test(args):
-    from symnet.symbol_resnet import get_resnet_test
-    args.img_pixel_means = (0.0, 0.0, 0.0)
-    args.img_pixel_stds = (1.0, 1.0, 1.0)
-    args.rpn_feat_stride = 16
-    args.rcnn_feat_stride = 16
-    args.rcnn_pooled_size = (14, 14)
-    return get_resnet_test(anchor_scales=args.rpn_anchor_scales, anchor_ratios=args.rpn_anchor_ratios,
-                           rpn_feature_stride=args.rpn_feat_stride, rpn_pre_topk=args.rpn_pre_nms_topk,
-                           rpn_post_topk=args.rpn_post_nms_topk, rpn_nms_thresh=args.rpn_nms_thresh,
-                           rpn_min_size=args.rpn_min_size,
-                           num_classes=args.rcnn_num_classes, rcnn_feature_stride=args.rcnn_feat_stride,
-                           rcnn_pooled_size=args.rcnn_pooled_size, rcnn_batch_size=args.rcnn_batch_size,
-                           units=(3, 4, 23, 3), filter_list=(256, 512, 1024, 2048))
-
-def get_network(network, args):
-    networks = {
-        'vgg16': get_vgg16_test,
-        'resnet50': get_resnet50_test,
-        'resnet101': get_resnet101_test
-    }
-    if network not in networks:
-        raise ValueError("network {} not supported".format(network))
-    return networks[network](args)
-
-
 def main():
     args = parse_args()
     class_names = get_dataset_demo(args.dataset, args)
-    sym = get_network(args.network, args)
+    sym = get_network_test(args.network, args)
     demo_net(sym, class_names, args)
